move normalizeRepeat to analysis stage
