+++ fe/fe-core/src/main/java/org/apache/doris/nereids/trees/expressions/Cast.java
@@ -62,12 +62,12 @@ public class Cast extends Expression implements UnaryExpression {
 
     @Override
     public String toSql() throws UnboundException {
-        return "CAST(" + child().toSql() + " AS " + targetType + ")";
+        return "cast(" + child().toSql() + " as " + targetType + ")";
     }
 
     @Override
     public String toString() {
-        return "CAST(" + child() + " AS " + targetType + ")";
+        return "cast(" + child() + " as " + targetType + ")";
     }
 
     @Override
